Add functionality to admin invalidations
Currently invalidating signatures is an ad-hoc task but with increasing
amounts of fraudulent signatures we need to make it an admin task.

This commit adds an invalidation model that tracks what are the reasons
for the invalidation and what are its criteria. It processes them in a
background job so that a sysadmin can track the progress of a large
invalidation and cancel it if necessary. You can also preview the number
of signatures that will be invalidated.